move functions to Zaps.ts
